requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
arrow==0.14.7
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.12.1
beautifulsoup4==4.12.3
billiard==3.6.4.0
boto3==1.20.5
botocore==1.23.54
celery==5.2.2
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
dj-rest-auth==5.1.0
Django==5.1
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.26.1
django-admin-numeric-filter==0.1.6
django-admin-rangefilter==0.8.5
django-admin-tools==0.9.2
django-allauth==0.56.0
django-appconf==1.0.6
django-background-tasks==1.2.8
django-braces==1.15.0
django-colorfield==0.10.1
django-cors-headers==4.4.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-debug-toolbar==4.3.0
django-encrypted-json-fields==1.0.4
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-filter==24.3
django-google-maps==0.13.0
django-grappelli==3.0.3
django-ical==1.8.3
django-image-uploader-widget==1.0.0
django-imagekit==5.0.0
django-import-export==2.8.0
django-ipware==4.0.2
django-jazzmin==3.0.0
django-json-widget==1.1.1
django-localflavor==3.1
django-map-widgets==0.4.1
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-oauth-toolkit==2.3.0
django-override-autonow==0.0.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-pivot==1.9.0
django-recurrence==1.11.1
django-redis==5.2.0
django-request==1.6.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-reversion==5.0.1
django-select2==7.10.0
django-sendgrid-v5==1.1.1
django-ses==2.3.1
django-simple-history==3.1.1
django-storages==1.12.3
django-uuid-upload-path==1.0.0
django-vectortiles==0.2.0
django-versatileimagefield==3.1
django4-background-tasks==1.2.7
djangogrpcframework==0.2.1
djangorestframework==3.15.2
djangorestframework-gis==1.0
dnspython==2.6.1
docutils==0.15.2
drf-extensions==0.7.1
drf-nested-routers==0.93.4
drf-recaptcha==2.2.2
drf-spectacular==0.26.4
drf-writable-nested==0.6.3
drf-yasg==1.20.0
easy-thumbnails==2.8.1
elasticsearch==7.15.0
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.1
fiona==1.9.6
fonttools==4.50.0
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
GeoAlchemy2==0.14.6
geocoder==1.38.1
geoip2==4.6.0
geojson-rewind==1.0.3
geopandas==0.11.1
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
icalendar==5.0.12
ics==0.7
idna==3.8
importlib_resources==6.4.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.3
jiwer==3.0.3
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.6
kiwisolver==1.4.5
kombu==5.3.5
libmagic==1.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
maxminddb==2.6.0
mdurl==0.1.2
mercadopago==2.2.0
mercantile==1.2.1
model-bakery==1.6.0
mortgage==1.0.5
multidict==6.0.5
nameparser==1.1.3
namesparser==1.0.0
natsort==8.3.1
numerize==0.12
numpy==2.1.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.9
orjson==3.10.0
oscrypto==1.3.0
packaging==23.2
pandas==2.2.2
phonenumbers==8.13.0
pilkit==3.0
Pillow==9.4.0
pip-system-certs==4.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.21.7
psycopg2-binary==2.9.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==4.3.1
pypng==0.20220715.0
pyproj==3.6.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.1
python-stdnum==1.18
python3-openid==3.2.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
qrcode==7.4.2
rapidfuzz==3.8.1
ratelim==0.1.6
redis==3.5.3
referencing==0.30.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.10.2
Rtree==1.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.7
s3transfer==0.5.2
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.11.2
semantic-version==2.10.0
sendgrid==6.8.1
serpy==0.3.1
Shapely==1.8.5.post1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
SQLAlchemy==2.0.27
sqlalchemy-filters==0.3.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
starlette==0.37.2
stripe==5.0.0
svglib==1.5.1
tablib==3.6.0
TatSu==5.8.3
text-unidecode==1.3
threadpoolctl==3.4.0
tinycss2==1.3.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.62.3
twilio==9.2.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.20
uuid==1.30
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
wrapt==1.16.0
xhtml2pdf==0.2.16
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.7
